refactor(hub-common): leverage dynamic imports
affects: @esri/hub-common
